@@ -11,7 +11,7 @@
   },
   "main": "index.js",
   "engines": {
-    "node": ">=7.6.0"
+    "node": ">=7.8.0"
   },
   "scripts": {
     "test": "standard && nsp check && ava",
@@ -21,7 +21,7 @@
     "setup": "npm install",
     "standard-fix": "standard --fix",
     "start": "micro",
-    "now-deploy": "npm test && now --npm -e GITHUB_USER=@micro-pr-github-user -e GITHUB_TOKEN=@micro-pr-github-token && now alias"
+    "now-deploy": "npm test && now --npm -E production.env && now alias"
   },
   "keywords": [
     "microservice",
@@ -45,7 +45,7 @@
     "type": "npm"
   },
   "devDependencies": {
-    "ava": "0.18.2",
+    "ava": "0.19.0",
     "coveralls": "2.13.0",
     "nsp": "2.6.3",
     "nyc": "10.2.0",
