@@ -3,9 +3,10 @@
 const config = require('../config')
 
 module.exports = data => {
-  const auth = 'Basic ' + new Buffer(config.GITHUB_USER + ':' + config.GITHUB_TOKEN).toString('base64')
+  const auth = 'Basic ' + Buffer.from(config.GITHUB_USER + ':' + config.GITHUB_TOKEN).toString('base64')
   const issue = {
     assignees: [
+      'CoderGard',
       'cybermelmac'
     ],
     labels: [
